backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
groq==0.13.1
deepgram-sdk>=3.11.0,<4
sqlalchemy==2.0.36
aiosqlite==0.20.0
python-dotenv==1.0.1
pydantic-settings==2.6.1
pypdf>=5.0.0
python-docx>=1.1.0
numpy>=1.26.0
sentence-transformers>=3.0.0
